requirement.txt
alembic==1.4.2
APScheduler==3.6.3
bcrypt==3.1.7
Bootstrap-Flask==1.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
ConfigArgParse==1.2
cryptography==2.9
DingtalkChatbot==1.5.0
dominate==2.5.1
Fabric3==1.14.post1
Flask==1.1.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gevent==1.5a3
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
grpcio==1.28.1
grpcio-tools==1.28.1
idna==2.9
invoke==1.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
locust==0.0
locust-plugins==0.0.42
locust-swarm==0.0.41
locustio==0.14.5
Mako==1.1.2
MarkupSafe==1.1.1
msgpack==1.0.0
numpy==1.18.3
pandas==1.0.3
paramiko==2.7.1
protobuf==3.11.3
psutil==5.7.0
psycogreen==1.0.2
psycopg2==2.8.4
psycopg2-binary==2.8.5
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
pyzmq==19.0.0
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
tzlocal==2.0.0
urllib3==1.25.8
visitor==0.1.3
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.2.1
